requirements.txt
flask~=3.1.2
flask-sqlalchemy
pymysql
google-generativeai
fpdf~=1.7.2
python-dotenv~=1.2.1
flask-login~=0.6.3
flask-cors~=6.0.1

protobuf~=5.29.5
PyJWT~=2.10.1